[4283] Allow set in mangosd.conf step of skinning/minning skill grow slowdown or disable slowdown ( SkillChance.MiningSteps SkillChance.SkinningSteps ). Patch provided by MoHax357. Fixed: crash at non-existed map loading after last map loading changes. Fixed: possible crash and remove redundant sql queries in .wp command code. Required lot more fixing but later. Fixed float to int32 warning conversion.

+++ src/game/Unit.cpp
@@ -4211,7 +4211,7 @@ void Unit::HandleProcTriggerSpell(Unit *pVictim, uint32 damage, Aura* triggeredB
         // Shamanistic Rage triggered spell
         case 30824:
         {
-            int32 SRBasePoints0 = GetTotalAttackPowerValue(BASE_ATTACK)*triggeredByAura->GetModifier()->m_amount/100 -1;
+            int32 SRBasePoints0 = int32(GetTotalAttackPowerValue(BASE_ATTACK)*triggeredByAura->GetModifier()->m_amount/100) -1;
             CastCustomSpell(this, 30824, &SRBasePoints0, NULL, NULL, true, NULL, triggeredByAura);
             return;
         }
